reorganised types and constants
