+++ src/constants/ribocode.ts
@@ -1,11 +1,14 @@
 /**
- * Types and constants for molecule modes in Ribocode.
+ * Ribocode constants used across the application, including viewer keys and alignment states.
  * 
  * Copyright (c) 2024-now Ribocode contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Andy Turner <agdturner@gmail.com>
  */
-export type MoleculeMode = 'AlignedTo' | 'Aligned' | 'ReAligned';
+import type { ViewerKey, MoleculeMode } from '../types/ribocode';
+
+export const A: ViewerKey = 'A';
+export const B: ViewerKey = 'B';
 export const AlignedTo: MoleculeMode = 'AlignedTo';
 export const Aligned: MoleculeMode = 'Aligned';
 export const ReAligned: MoleculeMode = 'ReAligned';
