+++ src/types/ribocode.ts
@@ -1,0 +1,9 @@
+/**
+ * Types for Ribocode.
+ * 
+ * Copyright (c) 2024-now Ribocode contributors, licensed under MIT, See LICENSE file for more info.
+ *
+ * @author Andy Turner <agdturner@gmail.com>
+ */
+export type ViewerKey = "A" | "B";
+export type MoleculeMode = 'AlignedTo' | 'Aligned' | 'ReAligned';
